move resolver transform into compat
